Инициализация сторов
