+++ frontend/src/store/cartStore.js
@@ -1,0 +1,3 @@
+import { defineStore } from "pinia";
+
+export const useCartStore = defineStore("cart", () => {});
