+++ frontend/src/store/dataStore.js
@@ -1,0 +1,3 @@
+import { defineStore } from "pinia";
+
+export const useDataStore = defineStore("data", () => {});
