+++ frontend/src/store/profileStore.js
@@ -1,0 +1,3 @@
+import { defineStore } from "pinia";
+
+export const useProfileStore = defineStore("profile", () => {});
