+++ frontend/src/views/MainView.vue
@@ -40,39 +40,21 @@
 
 <script setup>
 import CustomTitle from "@/common/components/CustomTitle.vue";
-import { ref } from "vue";
 import DoughConstructor from "@/modules/Constructor/DoughConstructor.vue";
 import DiameterConstructor from "@/modules/Constructor/DiameterConstructor.vue";
 import IngredientConstructor from "@/modules/Constructor/IngredientConstructor.vue";
 import PizzaConstructor from "@/modules/Constructor/PizzaConstructor.vue";
 import HeaderLayout from "@/layouts/HeaderLayout.vue";
+import { usePizzaStore } from "@/store/pizzaStore";
+import { storeToRefs } from "pinia";
 
-const pizzaDough = ref("light");
-const pizzaDiameter = ref("small");
-const pizzaSauce = ref("tomato");
-const pizzaIngredients = ref({
-  mushrooms: 0,
-  cheddar: 0,
-  salami: 0,
-  ham: 0,
-  ananas: 0,
-  bacon: 0,
-  onion: 0,
-  chile: 0,
-  jalapeno: 0,
-  olives: 0,
-  tomatoes: 0,
-  salmon: 0,
-  mozzarella: 0,
-  parmesan: 0,
-  blue_cheese: 0,
-});
+const { pizzaIngredients, pizzaDough, pizzaDiameter, pizzaSauce } = storeToRefs(
+  usePizzaStore()
+);
 
 const dropHandler = (transferData) => {
   pizzaIngredients.value[transferData.value]++;
 };
-
-//:-)
 </script>
 
 <style lang="scss">
